Fix Modal Open Bug
Since the modal body was wrapped in the same div
tag as the buttons (a tag that was the about the size of the buttons)
the modal body tried to open within the confines of that div tag causing
the bug. Wrap the modal buttons in one tag and leave the bodies of the
modals as a part of the overall div wrapper, giving them room to expand.